EDU-357 feat: 트랜잭션 함수 분리 및 문제 관련 스크립트 작성
